+++ examples/src/server/kinova.cpp
@@ -39,6 +39,7 @@ int main(int argc, char* argv[]) {
 
   /// launch raisim server
   raisim::RaisimServer server(&world);
+  server.setMap("simple");
   server.launchServer();
   server.focusOn(kinova);