+++ examples/src/server/visualObjects.cpp
@@ -32,7 +32,7 @@ int main(int argc, char* argv[]) {
   auto visSphere = server.addVisualSphere("v_sphere", 1.0, 1, 1, 1, 1);
   auto visBox = server.addVisualBox("v_box", 1, 1, 1, 1, 1, 1, 1);
   auto visCylinder = server.addVisualCylinder("v_cylinder", 1, 1, 0, 1, 0, 1);
-  auto visCapsule = server.addVisualCapsule("Cylinder", 1, 0.5, 0, 0, 1, 1);
+  auto visCapsule = server.addVisualCapsule("v_capsule", 1, 0.5, 0, 0, 1, 1);
   auto mesh = world.addMesh(binaryPath.getDirectory() + "/rsc/monkey/monkey.obj", 1, inertia, com);
   auto visMesh = server.addVisualMesh("v_mesh", binaryPath.getDirectory() + "/rsc/monkey/monkey.obj");
   auto anymalB = server.addVisualArticulatedSystem("v_anymal", binaryPath.getDirectory() + "/rsc/anymal/urdf/anymal.urdf");